Improve scraper with proxies. Add bit.io database connection WIP.

+++ potential-robot/scraper/allrecipe.py
@@ -9,39 +9,22 @@
 # -----------------------------------------------------------------------------
 
 
-import string
-import os
-import pandas as pd
-from bs4 import BeautifulSoup
-import requests
 import json
-import hashlib
+import os
+import string
 
+import pandas as pd
+import requests
+from bs4 import BeautifulSoup
+
+import helpers
 
 # -----------------------------------------------------------------------------
 # Hashing
 # -----------------------------------------------------------------------------
 
-def generate_dict_md5hash(input_dict):
-    """Generate an md5 hash for a dictionary"""
-    # # Remove an existing hash if there is one
-    # if input_dict.get('md5'):
-    #     input_dict.pop('md5')
-
-    # To generate a hash for input_dict it
-    # has to be converted to text form.
-    # Let's use json to help us.
-    input_dict_text = json.dumps(input_dict).encode('utf-8')
-    input_dict_hash = hashlib.md5(input_dict_text).hexdigest()
-
-    return input_dict_hash
-
-def insert_md5hash_to_dict(input_dict, hash):
-    input_dict['md5'] = hash
-    return input_dict
-
 def check_recipe_md5hash(input_dict):
-    new_md5 = generate_dict_md5hash(input_dict)
+    new_md5 = helpers.generate_dict_md5hash(input_dict)
     # Try to fetch the md5 hash from the database.
     # If a previous hash exists then compare both hashes.
 
@@ -208,8 +191,8 @@ def scrape_one_allrecipe(url: str) -> list:
     }
 
     # Generate a md5 hash for the recipe
-    hash = generate_dict_md5hash(recipe)
-    recipe = insert_md5hash_to_dict(recipe, hash)
+    hash = helpers.generate_dict_md5hash(recipe)
+    recipe = helpers.insert_md5hash_to_dict(recipe, hash)
 
     return recipe
 
@@ -223,5 +206,6 @@ def save_recipe_json(json_data):
     # Store the json file
     save_path = os.path.join('/', 'mnt', 'data_projects', 'potential-robot',
                              'temp', filename)
-    with open(save_path, 'w') as j:
-        json.dump(json_data, j, indent=4)
+    if not os.path.exists(save_path):
+        with open(save_path, 'w') as j:
+            json.dump(json_data, j, indent=4)
